test: stabilize provider session CLI coverage
